Move useExistingTagEffect to new hooks directory.
